+++ Contracts/Responses/Client/DeleteClientResponse.cs
@@ -2,15 +2,15 @@
 {
     public class DeleteClientResponse(
         string name,
-        float weight,
-        float height,
+        double weight,
+        double height,
         string email)
     {
         public string Name { get; } = name;
 
-        public float Weight { get; } = weight;
+        public double Weight { get; } = weight;
 
-        public float Height { get; } = height;
+        public double Height { get; } = height;
 
         public string Email { get; } = email;
     }
